All completed code from work hours. Test push

+++ pizzabox.Domain/Abstracts/A_Component.cs
@@ -3,6 +3,6 @@ namespace pizzabox.Domain.Abstracts
    public class A_Component
    {
       public string Name { get; set; }
-      public decimal Price { get; set; }
+      public double Price { get; set; }
    }
 }
